Dedup grant context text via Grant model method
